// src/interfaces/Fornecedor.ts
export interface FornecedorProps {
    id?: number;
    name?: string;
    cadastrado?: Date | string;
    atualizado?: Date | string;
    cnpj?: string;
    telefone?: string;
    numero?: string;
    email?: string;
    cep?: string;
    estado?: string;
    logradouro?: string;
    cidade?: string;
    pais?: string;
}